restructured app, added flask-migrate and created db
