+++ start-dev-env.sh
@@ -23,6 +23,7 @@ do
         -l|--local)
         # pull environmental variables for config. -s denotes as shell format. will require login if you aren't logged in to cli.
         # sed replaces "'" with nothing
+        export FLASK_APP="phish_bot.py"
         export APP_SETTINGS="config.DevelopmentConfig"
         export DATABASE_URL="postgresql:///phishbot_dev"
         shift
